[backend] add validation to FileController methods in API Gateway: implement getLecturers endpoint in the CoursesService

+++ HwProj.APIGateway/HwProj.APIGateway.API/Controllers/FilesController.cs
@@ -1,7 +1,9 @@
+using System.Linq;
 using System.Net;
 using System.Threading.Tasks;
 using HwProj.AuthService.Client;
 using HwProj.ContentService.Client;
+using HwProj.CoursesService.Client;
 using HwProj.Models.ContentService.DTO;
 using HwProj.Models.Roles;
 using Microsoft.AspNetCore.Authorization;
@@ -15,11 +17,14 @@ namespace HwProj.APIGateway.API.Controllers
     public class FilesController : AggregationController
     {
         private readonly IContentServiceClient _contentServiceClient;
+        private readonly ICoursesServiceClient _coursesServiceClient;
 
         public FilesController(IAuthServiceClient authServiceClient,
-            IContentServiceClient contentServiceClient) : base(authServiceClient)
+            IContentServiceClient contentServiceClient,
+            ICoursesServiceClient coursesServiceClient) : base(authServiceClient)
         {
             _contentServiceClient = contentServiceClient;
+            _coursesServiceClient = coursesServiceClient;
         }
 
         [HttpPost("upload")]
@@ -28,6 +33,10 @@ namespace HwProj.APIGateway.API.Controllers
         [ProducesResponseType(typeof(string[]), (int)HttpStatusCode.BadRequest)]
         public async Task<IActionResult> Upload([FromForm] UploadFileDTO uploadFileDto)
         {
+            var courseLecturersIds = await _coursesServiceClient.GetCourseLecturersIds(uploadFileDto.CourseId);
+            if (!courseLecturersIds.Contains(UserId))
+                return BadRequest("Пользователь с такой почтой не является преподавателем курса");
+
             var result = await _contentServiceClient.UploadFileAsync(uploadFileDto);
             return result.Succeeded ? Ok() as IActionResult : BadRequest(result.Errors);
         }
@@ -42,29 +51,28 @@ namespace HwProj.APIGateway.API.Controllers
                 ? Ok(result.Value) as IActionResult
                 : NotFound(result.Errors);
         }
-        
-        [HttpGet("homeworkFilesInfo/{courseId}/{homeworkId}")]
-        [ProducesResponseType(typeof(HomeworkFileInfoDTO[]), (int)HttpStatusCode.OK)]
-        public async Task<IActionResult> GetHomeworkFilesInfo(long courseId, long homeworkId)
+
+        [HttpGet("filesInfo/{courseId}")]
+        [ProducesResponseType(typeof(FileInfoDTO[]), (int)HttpStatusCode.OK)]
+        public async Task<IActionResult> GetFilesInfo(long courseId, [FromQuery] long? homeworkId = null)
         {
-            var filesInfo = await _contentServiceClient.GetHomeworkFilesInfo(courseId, homeworkId);
+            var filesInfo = await _contentServiceClient.GetFilesInfo(courseId, homeworkId);
             return Ok(filesInfo);
         }
 
-        [HttpGet("courseFilesInfo/{courseId}")]
-        [ProducesResponseType(typeof(CourseFileInfoDTO[]), (int)HttpStatusCode.OK)]
-        public async Task<IActionResult> GetCourseFilesInfo(long courseId)
-        {
-            var filesInfo = await _contentServiceClient.GetCourseFilesInfo(courseId);
-            return Ok(filesInfo);
-        }
-        
         [HttpDelete]
         [Authorize(Roles = Roles.LecturerRole)]
         [ProducesResponseType((int)HttpStatusCode.OK)]
         [ProducesResponseType(typeof(string[]), (int)HttpStatusCode.BadRequest)]
         public async Task<IActionResult> DeleteFile([FromQuery] string key)
         {
+            var courseIdResult = await _contentServiceClient.GetCourseIdFromKeyAsync(key);
+            if (!courseIdResult.Succeeded) return BadRequest(courseIdResult.Errors);
+
+            var courseLecturersIds = await _coursesServiceClient.GetCourseLecturersIds(courseIdResult.Value);
+            if (!courseLecturersIds.Contains(UserId))
+                return BadRequest("Пользователь с такой почтой не является преподавателем курса");
+
             var deletionResult = await _contentServiceClient.DeleteFileAsync(key);
             return deletionResult.Succeeded
                 ? Ok() as IActionResult
